+++ src/main/java/webserver/http/HttpRequest.java
@@ -1,7 +1,6 @@
-package http;
+package webserver.http;
 
-import static http.header.HeaderConst.*;
-import static http.statusline.ResponseLine.*;
+import static webserver.http.statusline.ResponseLine.*;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -10,9 +9,10 @@ import java.util.NoSuchElementException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import http.header.Header;
-import http.statusline.HttpMethod;
-import http.statusline.HttpVersion;
+import webserver.http.header.Header;
+import webserver.http.statusline.HttpMethod;
+import webserver.http.statusline.HttpVersion;
+import webserver.http.header.HeaderConst;
 
 public class HttpRequest {
 	private static final int KEY = 0;
@@ -51,12 +51,12 @@ public class HttpRequest {
 	}
 
 	private void parseHeader() throws IOException {
-		String line = HEADER_DELIMITER;
+		String line = HeaderConst.HEADER_DELIMITER;
 		String[] tokens;
 		while (!line.isBlank()) {
 			line = reader.readLine();
-			if (line.contains(HEADER_DELIMITER)) {
-				tokens = line.split(HEADER_DELIMITER);
+			if (line.contains(HeaderConst.HEADER_DELIMITER)) {
+				tokens = line.split(HeaderConst.HEADER_DELIMITER);
 				header.addHeader(tokens[KEY], tokens[VALUE]);
 			}
 		}
