+++ src/main/java/webserver/http/header/Header.java
@@ -1,6 +1,4 @@
-package http.header;
-
-import static http.header.HeaderConst.*;
+package webserver.http.header;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -12,9 +10,9 @@ public class Header {
 	public String getHeaderLines() {
 		StringBuilder stringBuilder = new StringBuilder();
 		for (String key : header.keySet()) {
-			stringBuilder.append(key + HEADER_DELIMITER + header.get(key) + CRLF);
+			stringBuilder.append(key + HeaderConst.HEADER_DELIMITER + header.get(key) + HeaderConst.CRLF);
 		}
-		stringBuilder.append(CRLF);
+		stringBuilder.append(HeaderConst.CRLF);
 		return stringBuilder.toString();
 	}
 
@@ -23,24 +21,29 @@ public class Header {
 	}
 
 	public boolean containsCookie() {
-		return header.containsKey(COOKIE);
+		return header.containsKey(HeaderConst.COOKIE);
 	}
 
 	public String getCookieValue(String cookieName) throws NoSuchElementException {
-		String cookieLines = header.get(COOKIE);
+		if (!header.containsKey(HeaderConst.COOKIE)) {
+			throw new NoSuchElementException("쿠키가 존재하지 않습니다.");
+		}
+		String cookieLines = header.get(HeaderConst.COOKIE);
 		for (String cookieLine : cookieLines.split("; ")) {
 			if (cookieLine.split("=")[0].equals(cookieName)) {
 				return cookieLine.split("=")[1];
 			}
 		}
+		System.out.println(HeaderConst.COOKIE);
+		System.out.println(cookieLines);
 		throw new NoSuchElementException("쿠키 이름에 해당하는 쿠키가 존재하지 않습니다.");
 	}
 
 	public boolean containsLength() {
-		return header.containsKey(CONTENT_LENGTH);
+		return header.containsKey(HeaderConst.CONTENT_LENGTH);
 	}
 
 	public int getContentLength() {
-		return Integer.parseInt(header.get(CONTENT_LENGTH));
+		return Integer.parseInt(header.get(HeaderConst.CONTENT_LENGTH));
 	}
 }
