+++ src/main/java/webserver/http/header/MimeType.java
@@ -1,4 +1,4 @@
-package http.header;
+package webserver.http.header;
 
 import java.util.Set;
