+++ src/main/java/webserver/http/statusline/StatusCode.java
@@ -1,4 +1,4 @@
-package http.statusline;
+package webserver.http.statusline;
 
 public enum StatusCode {
 	OK(200, "OK"),
