+++ src/main/java/webserver/session/SessionConst.java
@@ -1,4 +1,4 @@
-package session;
+package webserver.session;
 
 public interface SessionConst {
 	String sessionId = "LOGIN_MEMBER";
